# backend/docs/swagger.yaml
components:
  schemas:
    dto.CreateCurriculum:
      properties:
        group_id:
          type: string
        hours:
          minimum: 1
          type: integer
        lead_by:
          type: string
        report_type:
          $ref: '#/components/schemas/repositories.CurriculumReportType'
        semester:
          maximum: 12
          minimum: 1
          type: integer
        subject_id:
          type: string
      required:
      - group_id
      - hours
      - lead_by
      - report_type
      - semester
      - subject_id
      type: object
...
    dto.Curriculum:
      properties:
        group_id:
          type: string
        hours:
          type: integer
        id:
          type: string
        lead_by:
          type: string
        report_type:
          $ref: '#/components/schemas/repositories.CurriculumReportType'
        semester:
          type: integer
        subject_id:
          type: string
      type: object
    dto.Error:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
...
    dto.Login:
      properties:
        email:
          type: string
        password:
          minLength: 8
          type: string
      required:
      - email
      - password
      type: object
    dto.LoginResult:
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/dto.User'
      type: object
...
    dto.UpdateCurriculum:
      properties:
        group_id:
          type: string
        hours:
          minimum: 1
          type: integer
        lead_by:
          type: string
        report_type:
          $ref: '#/components/schemas/repositories.CurriculumReportType'
        semester:
          maximum: 12
          minimum: 1
          type: integer
        subject_id:
          type: string
      type: object
...
    dto.User:
      properties:
        created_at:
          type: string
        email:
          type: string
        id:
          type: string
        role:
          $ref: '#/components/schemas/repositories.UserRole'
        updated_at:
          type: string
      type: object
    repositories.CurriculumReportType:
      enum:
      - exam
      - test
      - diff_test
      - exam
      - test
      - diff_test
      - exam
      - test
      - diff_test
      type: string
      x-enum-varnames:
      - CurriculumReportTypeExam
      - CurriculumReportTypeTest
      - CurriculumReportTypeDiffTest
...
    repositories.UserRole:
      enum:
      - administrator
      - teacher
      - student
      - administrator
      - teacher
      - student
      - administrator
      - teacher
      - student
      type: string
      x-enum-varnames:
      - UserRoleAdministrator
      - UserRoleTeacher
      - UserRoleStudent
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: REST API образовательной системы Edutrack.
  title: Edutrack API
  version: "1.0"
openapi: 3.1.0
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.Login'
                description: Учётные данные
                summary: request
        description: Учётные данные
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.LoginResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      summary: Вход в аккаунт
      tags:
      - auth
  /curriculums:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.Curriculum'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить список учебных планов
      tags:
      - curriculums
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateCurriculum'
                description: Данные учебного плана
                summary: request
        description: Данные учебного плана
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Curriculum'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Создать учебный план
      tags:
      - curriculums
  /curriculums/{id}:
    delete:
      parameters:
      - description: Идентификатор учебного плана
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Удалить учебный план
      tags:
      - curriculums
    get:
      parameters:
      - description: Идентификатор учебного плана
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Curriculum'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить учебный план по идентификатору
      tags:
      - curriculums
    patch:
      parameters:
      - description: Идентификатор учебного плана
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateCurriculum'
                description: Данные учебного плана
                summary: request
        description: Данные учебного плана
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Curriculum'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Частично обновить учебный план
      tags:
      - curriculums
...
servers:
- url: /api